19 18 17 14 13 11 10 4 items are done as per SLL Test Sheet and Sale Order, Sale Invoice, Agreement Memo, Complaint Service PDF Download is working now

+++ app/src/main/java/com/lfsolutions/retail/model/service/ComplaintServiceRequest.kt
@@ -15,10 +15,10 @@ data class ComplaintServiceRequest(
 
     fun updatePriceAndQty() {
         var totalPrice = 0.0
-        var qty = 0
+        var qty = 0.0
         complaintServiceDetails.forEach {
             totalPrice += it.price
-            qty += it.qty?.toInt() ?: 0
+            qty += it.qty?.toDouble() ?: 0.0
         }
 
         complaintService?.totalQty = qty
